# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FinBridge API - Open Banking + Financial Inclusion platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessMessage {
  message: string;
}

export type UserKycTier = typeof UserKycTier[keyof typeof UserKycTier];


export const UserKycTier = {
  basic: 'basic',
  intermediate: 'intermediate',
  full: 'full',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bvn?: string | null;
  kycTier: UserKycTier;
  financialHealthScore: number;
  createdAt: string;
}

export interface UserUpdate {
  name?: string;
  phone?: string;
}

export type KycInputTier = typeof KycInputTier[keyof typeof KycInputTier];


export const KycInputTier = {
  intermediate: 'intermediate',
  full: 'full',
} as const;

export interface KycInput {
  tier: KycInputTier;
  /** @nullable */
  bvn?: string | null;
  /** @nullable */
  nin?: string | null;
  /** @nullable */
  address?: string | null;
}

export type BankAccountAccountType = typeof BankAccountAccountType[keyof typeof BankAccountAccountType];


export const BankAccountAccountType = {
  savings: 'savings',
  current: 'current',
  wallet: 'wallet',
} as const;

export interface BankAccount {
  id: number;
  bankName: string;
  accountNumber: string;
  accountType: BankAccountAccountType;
  balance: number;
  currency: string;
  isActive: boolean;
  connectedAt: string;
}

export type BankAccountInputAccountType = typeof BankAccountInputAccountType[keyof typeof BankAccountInputAccountType];


export const BankAccountInputAccountType = {
  savings: 'savings',
  current: 'current',
  wallet: 'wallet',
} as const;

export interface BankAccountInput {
  bankName: string;
  accountNumber: string;
  accountType: BankAccountInputAccountType;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  credit: 'credit',
  debit: 'debit',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  completed: 'completed',
  pending: 'pending',
  failed: 'failed',
} as const;

export interface Transaction {
  id: number;
  accountId: number;
  amount: number;
  type: TransactionType;
  category: string;
  description: string;
  date: string;
  status: TransactionStatus;
  /** @nullable */
  bankName?: string | null;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  credit: 'credit',
  debit: 'debit',
} as const;

export interface TransactionInput {
  accountId: number;
  amount: number;
  type: TransactionInputType;
  category: string;
  description: string;
  date?: string;
}

export interface TransactionList {
  transactions: Transaction[];
  total: number;
  limit: number;
  offset: number;
}

export interface CategorySpend {
  category: string;
  amount: number;
  count: number;
  percentage: number;
}

export interface TransactionAnalytics {
  totalSpend: number;
  totalIncome: number;
  byCategory: CategorySpend[];
  periodDays: number;
}

export interface SavingsGoal {
  id: number;
  name: string;
  targetAmount: number;
  currentAmount: number;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  autoDeductAmount?: number | null;
  autoDeductEnabled: boolean;
  createdAt: string;
}

export interface SavingsGoalInput {
  name: string;
  targetAmount: number;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  autoDeductAmount?: number | null;
  autoDeductEnabled?: boolean;
}

export interface SavingsGoalUpdate {
  name?: string;
  targetAmount?: number;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  autoDeductAmount?: number | null;
  autoDeductEnabled?: boolean;
}

export interface ContributionInput {
  amount: number;
}

export interface CommunityGroup {
  id: number;
  name: string;
  description: string;
  contributionAmount: number;
  cycleDays: number;
  totalPool: number;
  memberCount: number;
  /** @nullable */
  nextPayoutDate?: string | null;
  isMember: boolean;
  createdAt: string;
}

export interface CommunityGroupInput {
  name: string;
  description: string;
  contributionAmount: number;
  cycleDays: number;
}

export interface GroupMember {
  id: number;
  userId: number;
  name: string;
  rotationOrder: number;
  hasPaidCurrentCycle: boolean;
  joinedAt: string;
}

export interface CommunityGroupDetail {
  id: number;
  name: string;
  description: string;
  contributionAmount: number;
  cycleDays: number;
  totalPool: number;
  memberCount: number;
  /** @nullable */
  nextPayoutDate?: string | null;
  isMember: boolean;
  createdAt: string;
  members: GroupMember[];
}

export type FinancialInsightType = typeof FinancialInsightType[keyof typeof FinancialInsightType];


export const FinancialInsightType = {
  warning: 'warning',
  tip: 'tip',
  achievement: 'achievement',
  alert: 'alert',
} as const;

export type FinancialInsightImpact = typeof FinancialInsightImpact[keyof typeof FinancialInsightImpact];


export const FinancialInsightImpact = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface FinancialInsight {
  id: number;
  type: FinancialInsightType;
  title: string;
  description: string;
  recommendation: string;
  impact: FinancialInsightImpact;
  createdAt: string;
}

export interface DashboardSummary {
  totalBalance: number;
  monthlyIncome: number;
  monthlyExpenses: number;
  savingsRate: number;
  financialHealthScore: number;
  totalSavings: number;
  activeGroupsCount: number;
  accountsCount: number;
  recentTransactions: Transaction[];
  insightCount: number;
}

export type ListTransactionsParams = {
limit?: number;
offset?: number;
/**
 * @nullable
 */
accountId?: number | null;
/**
 * @nullable
 */
category?: string | null;
/**
 * @nullable
 */
type?: ListTransactionsType;
};

export type ListTransactionsType = typeof ListTransactionsType[keyof typeof ListTransactionsType] | null;


export const ListTransactionsType = {
  credit: 'credit',
  debit: 'debit',
} as const;