# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FinBridge API - Open Banking + Financial Inclusion platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all connected bank accounts
      responses:
        "200":
          description: List of bank accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
    post:
      operationId: connectAccount
      tags: [accounts]
      summary: Connect a new bank account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountInput"
      responses:
        "201":
          description: Account connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get a single bank account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bank account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
    delete:
      operationId: disconnectAccount
      tags: [accounts]
      summary: Disconnect a bank account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # ── TRANSACTIONS ──────────────────────────────────────────────────────────
...
  /community:
    get:
      operationId: listCommunityGroups
      tags: [community]
      summary: List all community savings groups
      responses:
        "200":
          description: Community groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityGroup"
    post:
      operationId: createCommunityGroup
      tags: [community]
      summary: Create a new Ajo/Esusu group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityGroupInput"
      responses:
        "201":
          description: Created community group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityGroup"
  /community/{id}:
    get:
      operationId: getCommunityGroup
      tags: [community]
      summary: Get a community group with members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Community group with members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityGroupDetail"
  /community/{id}/join:
    post:
      operationId: joinCommunityGroup
      tags: [community]
      summary: Join a community savings group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityGroup"
  /community/{id}/contribute:
    post:
      operationId: contributeToGroup
      tags: [community]
      summary: Make contribution to a community group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContributionInput"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityGroup"

  # ── INSIGHTS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ── USER ──────────────────────────────────────────────────────────────
...
    BankAccount:
      type: object
      properties:
        id:
          type: integer
        bankName:
          type: string
        accountNumber:
          type: string
        accountType:
          type: string
          enum: [savings, current, wallet]
        balance:
          type: number
        currency:
          type: string
        isActive:
          type: boolean
        connectedAt:
          type: string
          format: date-time
      required:
        - id
        - bankName
        - accountNumber
        - accountType
        - balance
        - currency
        - isActive
        - connectedAt
    BankAccountInput:
      type: object
      properties:
        bankName:
          type: string
        accountNumber:
          type: string
        accountType:
          type: string
          enum: [savings, current, wallet]
      required:
        - bankName
        - accountNumber
        - accountType

    # ── TRANSACTIONS ──────────────────────────────────────────────────────
...
    ContributionInput:
      type: object
      properties:
        amount:
          type: number
      required:
        - amount

    # ── COMMUNITY GROUPS ──────────────────────────────────────────────────
    CommunityGroup:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        contributionAmount:
          type: number
        cycleDays:
          type: integer
        totalPool:
          type: number
        memberCount:
          type: integer
        nextPayoutDate:
          type: ["string", "null"]
          format: date
        isMember:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - contributionAmount
        - cycleDays
        - totalPool
        - memberCount
        - isMember
        - createdAt
    CommunityGroupInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        contributionAmount:
          type: number
        cycleDays:
          type: integer
      required:
        - name
        - description
        - contributionAmount
        - cycleDays
    GroupMember:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        rotationOrder:
          type: integer
        hasPaidCurrentCycle:
          type: boolean
        joinedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - rotationOrder
        - hasPaidCurrentCycle
        - joinedAt
    CommunityGroupDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        contributionAmount:
          type: number
        cycleDays:
          type: integer
        totalPool:
          type: number
        memberCount:
          type: integer
        nextPayoutDate:
          type: ["string", "null"]
          format: date
        isMember:
          type: boolean
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/GroupMember"
      required:
        - id
        - name
        - description
        - contributionAmount
        - cycleDays
        - totalPool
        - memberCount
        - isMember
        - createdAt
        - members

    # ── INSIGHTS ──────────────────────────────────────────────────────────
...